ADDED: the file writing class
Created the file output class for writing out the results to a file. Also, added the creation of the directory for the output results. Finally, though no of them have been tested yet there is another method for writing out to the file. The second one is more abstract to allow for more columns to be written to.